+++ Common/src/main/java/fuzs/puzzleslib/impl/config/ConfigHolderImpl.java
@@ -98,7 +98,7 @@ public abstract class ConfigHolderImpl implements ConfigHolder.Builder, Freezabl
             }
         }
 
-        ProxyImpl.get().registerConfigurationScreenForHolder(this.modId);
+        ProxyImpl.get().registerConfigurationScreen(this.modId);
     }
 
     @Override